Explicitly typecast

+++ fmmax/fmm_matrices.py
@@ -41,8 +41,8 @@ def script_k_matrix_patterned(
     transverse_wavevectors: jnp.ndarray,
 ) -> jnp.ndarray:
     """Returns the patterned-layer script-k matrix from eq. 19 of [2012 Liu]."""
-    kx = transverse_wavevectors[..., 0]
-    ky = transverse_wavevectors[..., 1]
+    kx = transverse_wavevectors[..., 0].astype(z_permittivity_matrix.dtype)
+    ky = transverse_wavevectors[..., 1].astype(z_permittivity_matrix.dtype)
     z_inv_kx = jnp.linalg.solve(z_permittivity_matrix, utils.diag(kx))
     z_inv_ky = jnp.linalg.solve(z_permittivity_matrix, utils.diag(ky))
     return jnp.block(
@@ -58,8 +58,8 @@ def k_matrix_patterned(
     transverse_wavevectors: jnp.ndarray,
 ) -> jnp.ndarray:
     """Returns the k-matrix for patterned magnetic materials."""
-    kx = transverse_wavevectors[..., 0]
-    ky = transverse_wavevectors[..., 1]
+    kx = transverse_wavevectors[..., 0].astype(z_permeability_matrix.dtype)
+    ky = transverse_wavevectors[..., 1].astype(z_permeability_matrix.dtype)
     z_inv_kx = jnp.linalg.solve(z_permeability_matrix, utils.diag(kx))
     z_inv_ky = jnp.linalg.solve(z_permeability_matrix, utils.diag(ky))
     return jnp.block(
